fix: Refactor microservice codebase
