+++ middleware/errorHandler.js
@@ -1,0 +1,6 @@
+const errorHandler = (err, req, res, next) => {
+    console.error(err);
+    res.status(500).json({ error: err.message });
+  };
+  
+  export default errorHandler;
