+++ package.json
@@ -6,7 +6,7 @@
     "main": "index.js",
     "scripts": {
         "test": "NODE_OPTIONS='--experimental-vm-modules' jest",
-        "dev": "concurrently -k \"npx hardhat node\" \"wait-on http://localhost:8545/ && nodemon src/api/index.js\"",
+        "dev": "concurrently -k \"npx hardhat node\" \"wait-on http://localhost:8545/ && nodemon index.js\"",
         "lint": "eslint ."
     },
     "keywords": [],
@@ -14,7 +14,7 @@
     "license": "ISC",
     "devDependencies": {
         "@nomicfoundation/hardhat-toolbox": "^5.0.0",
-        "eslint": "^9.12.0",
+        "eslint": "^9.13.0",
         "eslint-plugin-jest": "^28.8.3",
         "hardhat": "^2.22.13",
         "jest": "^29.7.0",
