+++ utils/convertBigIntToString.js
@@ -1,0 +1,15 @@
+function convertBigIntToString(obj) {
+    if (typeof obj === 'bigint') {
+        return obj.toString();
+    } else if (Array.isArray(obj)) {
+        return obj.map(convertBigIntToString);
+    } else if (typeof obj === 'object' && obj !== null) {
+        return Object.fromEntries(
+            Object.entries(obj).map(([key, value]) => [key, convertBigIntToString(value)])
+        );
+    } else {
+        return obj;
+    }
+}
+
+export default convertBigIntToString;
